[REFACTOR] #34 면접자 지원페이지 style 파일 통합
